Do not try to move libnss and crontab files at install
Crontab files are no more useful.
libnss files are still redeployed on start for existing instances and
people having manually configured it.

+++ boot-fixpath.sh
@@ -10,7 +10,6 @@ set -x
 [ -f /etc/libnss-mysql.cfg ]      && rm -f /etc/libnss-mysql.cfg
 [ -f /etc/my.cnf ]                && rm -f /etc/my.cnf
 [ -f /etc/nsswitch.conf ]         && rm -f /etc/nsswitch.conf
-[ -f /etc/crontab ]               && rm -f /etc/crontab
 [ -d /etc/tuleap ]                && rm -rf /etc/tuleap
 [ -d /etc/httpd/conf ]            && rm -rf /etc/httpd/conf
 [ -d /etc/httpd/conf.d/tuleap-plugins ] && mv /etc/httpd/conf.d/tuleap-plugins /etc/httpd-conf.d-tuleap-plugins
@@ -27,11 +26,10 @@ cd /etc
 ln -s /data/etc/tuleap tuleap
 ln -s /data/etc/aliases aliases
 ln -s /data/etc/aliases.codendi aliases.codendi
-ln -s /data/etc/libnss-mysql-root.cfg libnss-mysql-root.cfg
-ln -s /data/etc/libnss-mysql.cfg libnss-mysql.cfg
+[ -f /data/etc/libnss-mysql-root.cfg ] && ln -s /data/etc/libnss-mysql-root.cfg libnss-mysql-root.cfg
+[ -f /data/etc/libnss-mysql.cfg ]      && ln -s /data/etc/libnss-mysql.cfg libnss-mysql.cfg
 ln -s /data/etc/my.cnf my.cnf
 ln -s /data/etc/nsswitch.conf nsswitch.conf
-ln -s /data/etc/crontab crontab
 
 cd /etc/logrotate.d
 ln -s /data/etc/logrotate.d/httpd httpd
